third_party/blink/web_tests/http/tests/misc/performance-entry-serializer-expected.txt

This is a testharness.js-based test.
FAIL PerformanceEntry subclasses should serialize all attributes assert_equals: PerformanceMark.detail expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
